+++ src/sparseml/evaluation/integrations_config.yaml
@@ -14,3 +14,4 @@
 # my_eval_integration: integrations/my_eval_integration.py
 # my_eval_integration2: integrations/my_eval_integration2.py
 
+perplexity: integrations/perplexity.py
